check topic exists before initialising consumer and move timeout to consumer manager.
