+++ lib/consumerManager.js
@@ -1,6 +1,6 @@
 var kafka = require('kafka-node'),
     config = require('../config'),
-    consumers = {},
+    consumers = new Object(),
     log = require('../logger'),
     logger = log.logger,
     uuid = require('uuid'),
@@ -94,11 +94,24 @@ var kafka = require('kafka-node'),
             logger.debug({ consumer: consumer.id }, 'lib/consumerManager : Consumer closed.');
             if (!!cb) cb();
         });
+    },
+
+    timeoutConsumers = function () {
+        logger.trace('lib/consumerManager : Looking for timed-out consumers.');
+        var timeoutTime = Date.now() - consumerTimeoutMs;
+        for (var i in consumers) {
+            var consumer = consumers[i];
+            if (consumer.lastPoll < timeoutTime) {
+                logger.trace({ consumer : consumer.id }, 'lib/consumerManager : Consumer timed-out.');
+                deleteConsumer(consumer);
+            }
+        }
     };
 
 module.exports = {
     get: get,
     add: add,
     createInstance: createConsumerInstance,
-    delete: deleteConsumer
+    delete: deleteConsumer,
+    timeout: timeoutConsumers
 };
